split action class + subclasses into separate files, added xp system
